requirements.txt
ansible==2.9.6
asn1crypto==1.3.0
bcrypt==3.1.7
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
enum34==1.1.10
idna==2.9
ipaddress==1.0.23
Jinja2==2.11.1
MarkupSafe==1.1.1
pan-python==0.16.0
pandevice==0.14.0
paramiko==2.7.1
psutil==5.7.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.5.1
PyNaCl==1.3.0
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
urllib3==1.25.8
xmltodict==0.12.0
